requirements.txt
py-cord==2.5.0
py-cord==2.5.0[voice]
python-dotenv==1.0.1
